An example of the use of the 2V3 packet protocol, written in Python
An example written in Python to illustrate the use of the 2V3 packet protocol to relay data between an MCU and a master device.
In this case, the Python script communicates with a temperature logger and logs the version of the software running in the logger, the serial numbers of the sensors and then logs the temperature measurements on a repetitive basis.  All logs are stored in a text file whose name is formed based on the date that the script is executed.